+++ NEWS.md
@@ -1,3 +1,9 @@
+# satres 1.1.0
+* Transformation vignette.
+* Functions to transform an object: `clip_bands()`, `select_bands()`, `merge_tiles()`.
+* Functions to query an object: `get_band_names()`, `get_spectral_band_names()`.
+* Change `satres()` parameter `only_bands` to `only_spectral_bands` to avoid ambiguity.
+
 # satres 1.0.0
 * Functionality for creating `satres` objects and exporting them.
 * Initial CRAN submission.